ai_engine/requirements.txt
PyPDF2
docx2txt
python-pptx
faiss-cpu
numpy
google-generativeai
wikipedia
requests
python-dotenv
